models fixed expect for war log

+++ src/models/language.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct Language {
     pub name: String,
-    pub id: u16,
+    pub id: u32,
     #[serde(rename = "languageCode")]
     pub language_code: String,
 }
